Add dynamic Cite page for future Incidents (#1751)
* Add dynamic Cite page for future Incidents

* Remove console.log

* Add test to `createCitationPages`

* Refactor cite and dynamic cite pages

* Remove `DEV_SSR` flag and create a `/pages/cite/[id].js` page for dynamic Incidents

* Fix test

* Rollback DEV_SSR: true setting
